Delete a company user who created a company is the onlyone allowed to delete the company when a company is deleted all accounts associated to that company are also deleted

+++ backend/app/utils/companyutils.py
@@ -96,15 +96,18 @@ def update_companyinfo(company: Company, user:User, data: dict) -> Tuple[Union[s
         db.session.rollback()
         return (str(e), 400)
 
-def delete_company(company_id:str):
+def delete_company(company_id:str, user:User):
     """delete a company"""
     try:
-        company = get_company(company_id)
+        # company = get_company(company_id)
+        company = UserCompanyAssociation.query.filter_by(company_id=company_id).first()
         if not company:
-            return 'company was deleted', 404
+            return 'company not found', 400
+        if company.user_id != user.id:
+            return 'You are not allowed to delete this company', 404
         db.session.delete(company)
         db.session.commit()
-        return f'{is_deleted} company  deleted successfully', 204
+        return 'company  deleted successfully', 200
     except Exception as error:
         db.session.rollback()
         return str(error), 400
